+++ test/query/def_builders_test.dart
@@ -67,10 +67,7 @@ void main() {
   test('query def should produce query definition', () {
     var q = TestQuery();
 
-    // Since each EntityListView generates its own pageId, we need to get it from the built query
     final builtQuery = q.queryBuilder().build();
-    final listDef = builtQuery.views['list1'] as ListQueryDef;
-    final pageId = listDef.pageId;
 
     // ignore: unused_local_variable
     var expected = QueryDefBuilder('TestEntity')
@@ -81,21 +78,16 @@ void main() {
           ..val('view3')
           ..val('view4');
       })
-      ..list('TestListEntity', 'list1', ['attr3', 'attr4'], pageId, (qb) {
+      ..list('TestListEntity', 'list1', ['attr3', 'attr4'], (qb) {
         qb
           ..val('view5')
           ..val('view6');
-      });
+      }, limit: 100);
 
-    // TODO: ListQueryDefBuilder is missing the length param, so this test will always fail.
-    // Since ListQueryDefBuilder is defined in horda_core, this will be fixed in the next reverse pagination PR.
-    // So we won't have to publish an extra horda_core and therefore horda_server version.
-    // expect(builtQuery.toJson(), expected.build().toJson());
+    expect(builtQuery.toJson(), expected.build().toJson());
   });
 
   test('query def builder should produce correct json', () {
-    const testPageId = 'test-page-id';
-
     var def = QueryDefBuilder('TestEntity')
       ..val('view1')
       ..val('view2')
@@ -104,7 +96,7 @@ void main() {
           ..val('view3')
           ..val('view4');
       })
-      ..list('TestListEntity', 'list1', ['attr3', 'attr4'], testPageId, (qb) {
+      ..list('TestListEntity', 'list1', ['attr3', 'attr4'], (qb) {
         qb
           ..val('view5')
           ..val('view6');
@@ -136,7 +128,9 @@ void main() {
             },
           },
           'attrs': ['attr3', 'attr4'],
-          'pageId': testPageId,
+          'startAfter': '',
+          'endBefore': '',
+          'limit': 0,
         },
       },
     });
@@ -169,6 +163,9 @@ void main() {
             },
           },
           'attrs': ['attr3', 'attr4'],
+          'startAfter': '',
+          'endBefore': '',
+          'limit': 0,
         },
       },
     };
@@ -204,6 +201,9 @@ void main() {
               'view6': {'type': 'val'},
             },
           },
+          'startAfter': '',
+          'endBefore': '',
+          'limit': 0,
         },
       },
     };
@@ -218,8 +218,6 @@ void main() {
   });
 
   test('query definition builder should build the right definition', () {
-    const testPageId = 'test-page-id';
-
     var qb = QueryDefBuilder('TestEntity')
       ..val('view11')
       ..val('view12')
@@ -228,7 +226,7 @@ void main() {
           ..val('view21')
           ..val('view22');
       })
-      ..list('TestListEntity', 'list1', ['attr1', 'attr2'], testPageId, (qb) {
+      ..list('TestListEntity', 'list1', ['attr1', 'attr2'], (qb) {
         qb
           ..val('view31')
           ..val('view32');
@@ -279,8 +277,6 @@ void main() {
   });
 
   test('query def builder should produce json', () {
-    const testPageId = 'test-page-id';
-
     var qb = QueryDefBuilder('TestEntity')
       ..val('view11')
       ..val('view12')
@@ -289,7 +285,7 @@ void main() {
           ..val('view21')
           ..val('view22');
       })
-      ..list('TestListEntity', 'list1', [], testPageId, (qb) {
+      ..list('TestListEntity', 'list1', [], (qb) {
         qb
           ..val('view31')
           ..val('view32');
@@ -321,9 +317,67 @@ void main() {
               'view32': {'type': 'val'},
             },
           },
-          'pageId': testPageId,
+          'startAfter': '',
+          'endBefore': '',
+          'limit': 0,
         },
       },
+    });
+  });
+
+  group('Pagination queryDefLimit', () {
+    test(
+      'ForwardPagination with positive limit should result in positive query def limit',
+      () {
+        final listView = EntityListView(
+          'testList',
+          query: TestListQuery(),
+          pagination: ForwardPagination(limitToFirst: 50),
+        );
+
+        final queryDef = listView.queryBuilder().build() as ListQueryDef;
+
+        expect(queryDef.limit, 50);
+      },
+    );
+
+    test('ForwardPagination with default limit should result in 100', () {
+      final listView = EntityListView(
+        'testList',
+        query: TestListQuery(),
+        pagination: ForwardPagination(),
+      );
+
+      final queryDef = listView.queryBuilder().build() as ListQueryDef;
+
+      expect(queryDef.limit, 100);
+    });
+
+    test(
+      'ReversePagination with positive limit should result in negative query def limit',
+      () {
+        final listView = EntityListView(
+          'testList',
+          query: TestListQuery(),
+          pagination: ReversePagination(limitToLast: 50),
+        );
+
+        final queryDef = listView.queryBuilder().build() as ListQueryDef;
+
+        expect(queryDef.limit, -50);
+      },
+    );
+
+    test('ReversePagination with default limit should result in -100', () {
+      final listView = EntityListView(
+        'testList',
+        query: TestListQuery(),
+        pagination: ReversePagination(),
+      );
+
+      final queryDef = listView.queryBuilder().build() as ListQueryDef;
+
+      expect(queryDef.limit, -100);
     });
   });
 }